+++ src/app/events/new/page.tsx
@@ -2,7 +2,7 @@ import CreateEventForm from "@/components/forms/CreateEventForm.tsx";
 
 export default function NewEventPage(){
     return (
-        <main className="min-h-screen px-6 py-16">
+        <main className="min-h-screen  px-6 py-16">
             <CreateEventForm/>
         </main>
     )